+++ src/converter/tree.hpp
@@ -42,7 +42,7 @@ struct points_collection_t
 inline void points_data_initialize(points_collection_t &to_init, const storage_header_t &header)
 {
   to_init.point_count = header.public_header.point_count;
-  to_init.data.emplace_back(header.input_id, offset_t(0), point_count_t(header.public_header.point_count));
+  to_init.data.emplace_back(header.input_id, offset_t(0), point_count_t(uint32_t(header.public_header.point_count)));
   to_init.max = header.morton_max;
   to_init.min = header.morton_min;
   to_init.min_lod = header.lod_span;
@@ -73,7 +73,7 @@ inline void points_data_add(points_collection_t &dest, const storage_header_t &t
     points_data_initialize(dest, to_add);
     return;
   }
-  dest.data.emplace_back(to_add.input_id, offset_t(0), point_count_t(to_add.public_header.point_count));
+  dest.data.emplace_back(to_add.input_id, offset_t(0), point_count_t(uint32_t(to_add.public_header.point_count)));
   dest.point_count += to_add.public_header.point_count;
   if (to_add.morton_min < dest.min)
     dest.min = to_add.morton_min;
